Modify views associated with SystemDashboards
