+++ app/controllers/author_dashboards/repositories_controller.rb
@@ -1,5 +1,9 @@
 module AuthorDashboards
   class RepositoriesController < AuthorDashboards::ApplicationController
+    # Concern overrides #update action of the Administrate::Administrate::ApplicationController
+    # to be able to call background jobs
+    include DashboardRepositories
+
     # Override `create` action to add association to `current_author`
     # and to call background jobs
     def create
@@ -22,30 +26,6 @@ module AuthorDashboards
       else
         render :new, locals: {
           page: Administrate::Page::Form.new(dashboard, resource)
-        }, status: :unprocessable_entity
-      end
-    end
-
-    # Override `update` action to call background jobs
-    def update
-      former_name = requested_resource.name
-      former_branch = requested_resource.branch
-
-      if requested_resource.update(resource_params)
-        if former_name != requested_resource.name || former_branch != requested_resource.branch
-          old_directory = Rails.root.join('repos', requested_resource.author.github_username, former_name)
-          FileUtils.remove_dir(old_directory) if Dir.exist?(old_directory)
-          CloneGithubRepoJob.perform_async(requested_resource.id)
-        else
-          PullGithubRepoJob.perform_async(requested_resource.id)
-        end
-        redirect_to(
-          after_resource_updated_path(requested_resource),
-          notice: translate_with_resource('update.success')
-        )
-      else
-        render :edit, locals: {
-          page: Administrate::Page::Form.new(dashboard, requested_resource)
         }, status: :unprocessable_entity
       end
     end
